refactor: update imports after package restructuring
