+++ core/src/main/kotlin/io/qent/sona/core/settings/SettingsRepository.kt
@@ -1,4 +1,4 @@
-package io.qent.sona.core
+package io.qent.sona.core.settings
 
 interface SettingsRepository {
     suspend fun load(): Settings
