+++ src/main/kotlin/io/qent/sona/actions/CreateNewChatAction.kt
@@ -5,7 +5,7 @@ import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.components.service
 import io.qent.sona.PluginStateFlow
-import io.qent.sona.core.State
+import io.qent.sona.core.state.State
 
 class CreateNewChatAction : AnAction("Create", "Create new chat", AllIcons.General.Add) {
     override fun actionPerformed(e: AnActionEvent) {
