+++ src/main/kotlin/io/qent/sona/repositories/PluginFilePermissionsRepository.kt
@@ -2,7 +2,7 @@ package io.qent.sona.repositories
 
 import com.intellij.openapi.project.Project
 import io.qent.sona.config.SonaConfig
-import io.qent.sona.core.FilePermissionsRepository
+import io.qent.sona.core.permissions.FilePermissionsRepository
 
 class PluginFilePermissionsRepository(project: Project) : FilePermissionsRepository {
     private val root = project.basePath ?: "/"
